Fix: move tryReadJSON to util
