+++ tests/util.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+
 function extractRange(text) {
   const index = text.indexOf('|');
   return {
@@ -11,5 +13,15 @@ function drawRange(text, index) {
   return text.slice(0, index) + '|' + text.slice(index);
 }
 
+function tryReadJSON(path) {
+  let text = null;
+  try {
+    text = fs.readFileSync(path, 'utf8');
+  } catch (err) {
+    return;
+  }
+  return JSON.parse(text);
+}
+
 // https://github.com/avajs/ava/issues/1319
-module.exports = {extractRange, drawRange};
+module.exports = {extractRange, drawRange, tryReadJSON};
